defend_data_capture/requirements.txt
appdirs==1.4.4
asgiref==3.3.1
attrs==20.3.0
black==20.8b1
certifi==2020.12.5
cfgv==3.2.0
chardet==4.0.0
click==7.1.2
convertdate==2.3.2
defusedxml==0.6.0
diff-match-patch==20200713
distlib==0.3.1
django>=3.1.6
django-environ==0.4.5
django-import-export==2.5.0
django-reversion==3.0.9
django-staff-sso-client==3.1.0
django-webpack-loader==0.7.0
djangorestframework==3.12.2
et-xmlfile==1.0.1
factory-boy==3.2.0
Faker==6.0.0
filelock==3.0.12
hijri-converter==2.1.1
holidays==0.10.5.2
identify==1.5.13
idna==2.10
iniconfig==1.1.1
jdcal==1.4.1
korean-lunar-calendar==0.2.1
MarkupPy==1.14
mypy-extensions==0.4.3
nodeenv==1.5.0
numpy==1.20.1
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.6
packaging==20.9
pandas==1.2.3
pathspec==0.8.1
pluggy==0.13.1
pre-commit==2.10.0
psycopg2-binary==2.8.6
py==1.10.0
PyMeeus==0.5.11
pyparsing==2.4.7
pytest==6.2.2
pytest-django==4.1.0
python-dateutil==2.8.1
pytz==2020.5
PyYAML==5.4.1
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
sentry-sdk==0.20.3
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
text-unidecode==1.3
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.4
virtualenv==20.4.0
xlrd==2.0.1
xlwt==1.3.0
